+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Master
+
+- Removed semaphore lock macros `ESNewLock()`, `ESLock(lock)`, `ESUnlock(lock)`
+
 ## 3.18.1 (2019-08-20)
 
 - Added semaphore lock macros `ESNewLock()`, `ESLock(lock)`, `ESUnlock(lock)`